Fixes with bullets and player
Changed sizeof bullets so they actually look proper size. Also, I changed the buttons so if you hold down the bullet key it will shoot repeatly. The players flinching wasn't rendering when it got hit or took fall damage so I fixed that bug. Also touched up on the smoothness of the tilemap